search: use default input, refactor, simplify
